marketing/requirements.txt
openai
pandas
resend
pillow
langgraph
python-dotenv